requirements.txt
beautifulsoup4==4.11.1
bibtexparser==1.4.3
openai==1.63.2
pandas==2.2.3
Pillow==11.1.0
pymupdf==1.25.3
Requests==2.32.3
webcolors==24.11.1
